pyproject.toml
[tool.poetry]
name = "chisato"
version = "0.1.0"
description = "An osu! collection showcase and backup"
authors = ["Phawit Pornwattanakul <me@helloyeew.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
python-decouple = "^3.8"
django-admin-interface = "^0.24.2"
django-crispy-forms = "^2.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
crispy-bootstrap5 = "^0.7"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
requests = "^2.28.2"
boto3 = "^1.26.108"
pika = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
